fix: close atm-graft smoke qa findings
