+++ scripts/smoke/run_thorough.py
@@ -1,16 +1,20 @@
 #!/usr/bin/env python3
 from __future__ import annotations
 
-from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
+from dataclasses import field
 from datetime import datetime, timezone
-import os
 from pathlib import Path
 from typing import Any, Callable
 import json
 import subprocess
 import shutil
 import time
+
+from run_thorough_graft import enable_graft_config
+from run_thorough_graft import run_graft_lane
+from run_thorough_retry import verify_retry_evidence
+from run_thorough_shared_host import run_shared_host_lane
 
 
 @dataclass(frozen=True)
@@ -28,26 +32,13 @@ class ThoroughSmokeRuntime:
     pass_row: Callable[[Any, str], None]
     fail_row: Callable[..., None]
     failure_mentions: Callable[[Any, str], bool]
-    analyze_log_text: Callable[[str, list[str]], Any]
+    analyze_log_text: Callable[..., Any]
     stop_daemon: Callable[..., None]
     process_is_alive: Callable[[int], bool]
     team: str
     operator: str
     recipient: str
-
-
-def enable_graft_config(workspace_dir: Path) -> None:
-    config_path = workspace_dir / ".atm.toml"
-    config_text = config_path.read_text(encoding="utf-8")
-    if "[atm.graft]" in config_text:
-        return
-    graft_section = '\n[atm.graft]\nenabled = true\n'
-    config_path.write_text(config_text.rstrip() + graft_section, encoding="utf-8")
-
-
-def graft_smoke_example_path(root: Path) -> Path:
-    example_name = "smoke_same_host.exe" if os.name == "nt" else "smoke_same_host"
-    return root / "target" / "release" / "examples" / example_name
+    allowed_error_codes: list[str] = field(default_factory=list)
 
 
 def run_thorough(binary_sha: str, runtime: ThoroughSmokeRuntime) -> dict[str, object]:
@@ -417,161 +408,8 @@ def run_thorough(binary_sha: str, runtime: ThoroughSmokeRuntime) -> dict[str, ob
             )
             status = "failed"
 
-        ready_path = fixture.root / "graft-ready"
-        ready_path.unlink(missing_ok=True)
-        graft_env = runtime.smoke_env(fixture, identity=runtime.recipient, root=runtime.root)
-        graft_send_payload: dict[str, object] | None = None
-        graft_stdout = ""
-        graft_stderr = ""
-        graft_error: str | None = None
-        graft_payload: dict[str, object] | None = None
-        graft_process = subprocess.Popen(
-            [
-                str(graft_smoke_example_path(runtime.root)),
-                str(fixture.workspace_dir),
-                runtime.team,
-                runtime.recipient,
-                f"{runtime.operator}@{runtime.team}",
-                "thorough smoke graft requires ack",
-                runtime.operator,
-                str(ready_path),
-            ],
-            cwd=fixture.workspace_dir,
-            env=graft_env,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-        )
-        try:
-            ready_deadline = time.perf_counter() + 15.0
-            while not ready_path.exists():
-                if graft_process.poll() is not None:
-                    graft_stdout, graft_stderr = graft_process.communicate()
-                    graft_error = "atm-graft smoke host exited before reporting ready"
-                    break
-                if time.perf_counter() >= ready_deadline:
-                    graft_process.kill()
-                    graft_stdout, graft_stderr = graft_process.communicate()
-                    graft_error = "timed out waiting for atm-graft smoke host readiness"
-                    break
-                time.sleep(0.05)
-
-            if graft_error is None:
-                try:
-                    graft_send_payload = runtime.parse_json_output(
-                        runtime.run_atm(
-                            runtime.root,
-                            base_env,
-                            fixture.workspace_dir,
-                            "send",
-                            runtime.recipient,
-                            "thorough smoke graft requires ack",
-                            "--requires-ack",
-                            "--json",
-                        )
-                    )
-                    try:
-                        graft_stdout, graft_stderr = graft_process.communicate(timeout=20)
-                    except subprocess.TimeoutExpired:
-                        graft_process.kill()
-                        graft_stdout, graft_stderr = graft_process.communicate()
-                        graft_error = (
-                            "atm-graft smoke host timed out before completing the ICD flow"
-                        )
-                    if graft_error is None and graft_process.returncode == 0:
-                        graft_payload = json.loads(graft_stdout)
-                except Exception as exc:
-                    graft_error = str(exc)
-        finally:
-            if graft_process.poll() is None:
-                graft_process.kill()
-                graft_process.communicate()
-
-        if graft_error is not None or graft_process.returncode != 0 or graft_payload is None:
-            runtime.fail_row(
-                rows["GRAFT-001"],
-                observed=json.dumps(
-                    {
-                        "send": graft_send_payload,
-                        "graft_error": graft_error,
-                        "stdout": graft_stdout,
-                        "stderr": graft_stderr,
-                        "returncode": graft_process.returncode,
-                    },
-                    indent=2,
-                ),
-                expected="the atm-graft host registers, receives one advisory nudge, reads and acknowledges the nudged message, and sends one unary follow-up back to the CLI operator",
-                root_cause="the same-host atm-graft smoke host exited before completing the advisory plus unary ICD lane",
-                artifact="atm-graft smoke host stdout/stderr",
-                notes="same-host atm-graft advisory and unary ICD lane failed",
-            )
+        if not run_graft_lane(runtime, rows, fixture, base_env):
             status = "failed"
-        else:
-            ack_reply_read = runtime.parse_json_output(
-                runtime.run_atm(
-                    runtime.root,
-                    base_env,
-                    fixture.workspace_dir,
-                    "read",
-                    runtime.operator,
-                    "--team",
-                    runtime.team,
-                    "--all",
-                    "--message-id",
-                    str(graft_payload["ack_reply_message_id"]),
-                    "--json",
-                )
-            )
-            follow_up_read = runtime.parse_json_output(
-                runtime.run_atm(
-                    runtime.root,
-                    base_env,
-                    fixture.workspace_dir,
-                    "read",
-                    runtime.operator,
-                    "--team",
-                    runtime.team,
-                    "--all",
-                    "--message-id",
-                    str(graft_payload["follow_up_message_id"]),
-                    "--json",
-                )
-            )
-            graft_ok = (
-                graft_payload.get("status") == "passed"
-                and graft_payload.get("nudge_count") == 1
-                and graft_payload.get("nudge_from") == runtime.operator
-                and graft_payload.get("nudge_message_id") == str(graft_send_payload["message_id"])
-                and graft_payload.get("read_selected_message_id")
-                == str(graft_send_payload["message_id"])
-                and ack_reply_read.get("selected_message_id")
-                == str(graft_payload["ack_reply_message_id"])
-                and follow_up_read.get("selected_message_id")
-                == str(graft_payload["follow_up_message_id"])
-            )
-            if graft_ok:
-                runtime.pass_row(
-                    rows["GRAFT-001"],
-                    "a real atm-graft host registered, consumed the advisory nudge, read and acknowledged the nudged message, and sent a unary follow-up back to the CLI operator",
-                )
-            else:
-                runtime.fail_row(
-                    rows["GRAFT-001"],
-                    observed=json.dumps(
-                        {
-                            "send": graft_send_payload,
-                            "graft_payload": graft_payload,
-                            "ack_reply_read": ack_reply_read,
-                            "follow_up_read": follow_up_read,
-                        },
-                        indent=2,
-                    ),
-                    expected="the atm-graft host registers, receives one advisory nudge, reads and acknowledges the nudged message, and sends one unary follow-up back to the CLI operator",
-                    root_cause="the same-host atm-graft advisory and unary ICD lane diverged before the smoke runner could prove the retained CLI and graft surfaces share the accepted daemon contract",
-                    artifact="atm-graft smoke host JSON plus operator-side read outputs",
-                    notes="same-host atm-graft advisory and unary ICD lane failed",
-                )
-                status = "failed"
 
         if daemon_pid is not None:
             runtime.stop_daemon(
@@ -591,6 +429,7 @@ def run_thorough(binary_sha: str, runtime: ThoroughSmokeRuntime) -> dict[str, ob
                 '"outcome":"delivery_policy.ack_reply.delivered"',
                 '"action":"shutdown_completed"',
             ],
+            allowed_error_codes=runtime.allowed_error_codes,
         )
         analysis = {
             "passed": analysis_result.passed,
@@ -935,251 +774,20 @@ def run_thorough(binary_sha: str, runtime: ThoroughSmokeRuntime) -> dict[str, ob
             )
             status = "failed"
 
-        shared_host_fixture_pair = runtime.create_shared_host_fixture_pair(
-            prefix="z21s.",
-            team_name_a="z21-shared-a",
-            team_name_b="z21-shared-b",
-            operator_a="z21-shared-operator-a",
-            operator_b="z21-shared-operator-b",
-            recipient_a="z21-shared-recipient-a",
-            recipient_b="z21-shared-recipient-b",
+        shared_host_ok, shared_host_fixture_pair, shared_daemon_pid = run_shared_host_lane(
+            runtime, rows
         )
-        shared_a = shared_host_fixture_pair.workspace_a
-        shared_b = shared_host_fixture_pair.workspace_b
-        shared_env_a = runtime.smoke_env(shared_a, identity=shared_a.operator, root=runtime.root)
-        shared_env_b = runtime.smoke_env(shared_b, identity=shared_b.operator, root=runtime.root)
-        shared_doctor_a = runtime.parse_json_output(
-            runtime.run_atm(
-                runtime.root, shared_env_a, shared_a.workspace_dir, "doctor", "--json"
-            )
-        )
-        shared_doctor_b = runtime.parse_json_output(
-            runtime.run_atm(
-                runtime.root, shared_env_b, shared_b.workspace_dir, "doctor", "--json"
-            )
-        )
-        shared_pid_a = shared_doctor_a.get("runtime_status", {}).get("singleton_owner_pid")
-        shared_pid_b = shared_doctor_b.get("runtime_status", {}).get("singleton_owner_pid")
-        if shared_pid_a is not None:
-            shared_daemon_pid = int(shared_pid_a)
-        for fixture_item, env_item in ((shared_a, shared_env_a), (shared_b, shared_env_b)):
-            runtime.run_atm(
-                runtime.root,
-                env_item,
-                fixture_item.workspace_dir,
-                "teams",
-                "add-member",
-                fixture_item.team_name,
-                fixture_item.operator,
-                "--json",
-            )
-            runtime.run_atm(
-                runtime.root,
-                env_item,
-                fixture_item.workspace_dir,
-                "teams",
-                "add-member",
-                fixture_item.team_name,
-                fixture_item.recipient,
-                "--json",
-            )
-
-        def run_send(
-            fixture_item: Any,
-            env_item: dict[str, str],
-            body: str,
-        ) -> dict[str, object]:
-            target = f"{fixture_item.recipient}@{fixture_item.team_name}"
-            return runtime.parse_json_output(
-                runtime.run_atm(
-                    runtime.root,
-                    env_item,
-                    fixture_item.workspace_dir,
-                    "send",
-                    target,
-                    body,
-                    "--requires-ack",
-                    "--json",
-                )
-            )
-
-        with ThreadPoolExecutor(max_workers=2) as pool:
-            send_future_a = pool.submit(
-                run_send,
-                shared_a,
-                shared_env_a,
-                "shared-host message from workspace A",
-            )
-            send_future_b = pool.submit(
-                run_send,
-                shared_b,
-                shared_env_b,
-                "shared-host message from workspace B",
-            )
-            shared_send_a = send_future_a.result()
-            shared_send_b = send_future_b.result()
-
-        shared_message_id_a = str(shared_send_a["message_id"])
-        shared_message_id_b = str(shared_send_b["message_id"])
-
-        def read_and_ack(
-            fixture_item: Any,
-            env_item: dict[str, str],
-            message_id: str,
-            ack_body: str,
-        ) -> dict[str, object]:
-            read_payload = runtime.parse_json_output(
-                runtime.run_atm(
-                    runtime.root,
-                    env_item,
-                    fixture_item.workspace_dir,
-                    "read",
-                    fixture_item.recipient,
-                    "--team",
-                    fixture_item.team_name,
-                    "--all",
-                    "--message-id",
-                    message_id,
-                    "--json",
-                )
-            )
-            ack_payload = runtime.parse_json_output(
-                runtime.run_atm(
-                    runtime.root,
-                    env_item,
-                    fixture_item.workspace_dir,
-                    "ack",
-                    message_id,
-                    ack_body,
-                    "--team",
-                    fixture_item.team_name,
-                    "--as",
-                    fixture_item.recipient,
-                    "--json",
-                )
-            )
-            return {"read": read_payload, "ack": ack_payload}
-
-        with ThreadPoolExecutor(max_workers=2) as pool:
-            read_ack_future_a = pool.submit(
-                read_and_ack,
-                shared_a,
-                shared_env_a,
-                shared_message_id_a,
-                "shared-host ack A",
-            )
-            read_ack_future_b = pool.submit(
-                read_and_ack,
-                shared_b,
-                shared_env_b,
-                shared_message_id_b,
-                "shared-host ack B",
-            )
-            shared_read_ack_a = read_ack_future_a.result()
-            shared_read_ack_b = read_ack_future_b.result()
-
-        shared_list_a = runtime.parse_json_output(
-            runtime.run_atm(runtime.root, shared_env_a, shared_a.workspace_dir, "list", "--json")
-        )
-        shared_list_b = runtime.parse_json_output(
-            runtime.run_atm(runtime.root, shared_env_b, shared_b.workspace_dir, "list", "--json")
-        )
-        shared_log_snapshot_a = runtime.parse_json_output(
-            runtime.run_atm(
-                runtime.root, shared_env_a, shared_a.workspace_dir, "log", "snapshot", "--json"
-            )
-        )
-        shared_records_a = json.dumps(shared_list_a)
-        shared_records_b = json.dumps(shared_list_b)
-        shared_host_ok = (
-            shared_doctor_a.get("summary", {}).get("status") == "healthy"
-            and shared_doctor_b.get("summary", {}).get("status") == "healthy"
-            and shared_pid_a is not None
-            and shared_pid_a == shared_pid_b
-            and shared_send_a.get("outcome") == "sent"
-            and shared_send_b.get("outcome") == "sent"
-            and shared_read_ack_a["read"].get("selected_message_id") == shared_message_id_a
-            and shared_read_ack_b["read"].get("selected_message_id") == shared_message_id_b
-            and shared_read_ack_a["ack"].get("message_id") == shared_message_id_a
-            and shared_read_ack_b["ack"].get("message_id") == shared_message_id_b
-            and shared_message_id_b not in shared_records_a
-            and shared_message_id_a not in shared_records_b
-            and isinstance(shared_log_snapshot_a.get("records"), list)
-            and runtime.process_is_alive(int(shared_pid_a))
-        )
-        if shared_host_ok:
-            runtime.pass_row(
-                rows["PRR-001"],
-                "two workspaces with one shared ATM_HOME daemon/database/log root handled concurrent send/read/ack traffic without cross-workspace leakage",
-            )
-        else:
-            runtime.fail_row(
-                rows["PRR-001"],
-                observed=json.dumps(
-                    {
-                        "doctor_a": shared_doctor_a,
-                        "doctor_b": shared_doctor_b,
-                        "send_a": shared_send_a,
-                        "send_b": shared_send_b,
-                        "read_ack_a": shared_read_ack_a,
-                        "read_ack_b": shared_read_ack_b,
-                        "list_a": shared_list_a,
-                        "list_b": shared_list_b,
-                        "log_snapshot_a": shared_log_snapshot_a,
-                    },
-                    indent=2,
-                ),
-                expected="two or more workspaces share one host daemon/database/log root, concurrent send/read/ack succeeds, no cross-workspace message leakage occurs, and the shared daemon remains healthy",
-                root_cause="the shared-host same-daemon smoke lane diverged before proving the accepted multi-workspace topology",
-                artifact="shared-host doctor/send/read/ack/list/log snapshot outputs",
-                notes="shared-host multi-workspace smoke coverage failed",
-            )
+        if not shared_host_ok:
             status = "failed"
 
         if shared_daemon_pid is not None and runtime.process_is_alive(shared_daemon_pid):
             runtime.stop_daemon(
                 shared_daemon_pid,
-                shared_a.home_dir / ".atm" / "daemon" / "atm-daemon.sock",
+                shared_host_fixture_pair.workspace_a.home_dir / ".atm" / "daemon" / "atm-daemon.sock",
             )
             shared_daemon_pid = None
 
-        retry_outcomes = {
-            "initial_miss",
-            "retry_attempt",
-            "acquired",
-            "spawn_requested",
-            "publish_wait_started",
-            "publish_wait_continuing",
-            "connected",
-        }
-        observed_retry_outcomes = set()
-        for record in copied_log_snapshot.get("records", []):
-            if not isinstance(record, dict):
-                continue
-            message = str(record.get("message", ""))
-            marker = "with outcome "
-            if marker in message:
-                observed_retry_outcomes.add(message.split(marker, 1)[1].strip())
-        if retry_outcomes.issubset(observed_retry_outcomes):
-            runtime.pass_row(
-                rows["Z1-009"],
-                "copied-state log snapshot retained the expected retry-visible daemon lifecycle outcomes while the durable send/read path succeeded",
-            )
-        else:
-            runtime.fail_row(
-                rows["Z1-009"],
-                observed=json.dumps(
-                    {
-                        "observed_outcomes": sorted(observed_retry_outcomes),
-                        "records": copied_log_snapshot.get("records", []),
-                    },
-                    indent=2,
-                ),
-                expected="log snapshot includes initial_miss, retry_attempt, acquired, spawn_requested, publish_wait_started, publish_wait_continuing, and connected while the durable copied-state lane succeeds",
-                root_cause="retry-visible daemon lifecycle evidence was not preserved in the retained copied-state log snapshot",
-                artifact="copied-state log snapshot --json",
-                notes="retry-visible daemon/runtime evidence was incomplete",
-            )
+        if not verify_retry_evidence(runtime, rows, copied_log_snapshot):
             status = "failed"
     except Exception as exc:
         status = "failed"
